openvswitch: do not trigger build of libunwind when it's not available
The current form of "+libunwind" will cause it to be built if any of the
packages defined in this Makefile is enabled (y or m).  This is at the
moment by-design of the build system.

Libunwind does not support architectures like arc.  Use the conditional
select To avoid (libunwind) build failures like the following,

  checking for ELF helper width... configure: error: Unknown ELF target: arc
  make[3]: *** [Makefile:65: /data/openwrt/build_dir/target-arc_arc700_uClibc/
  libunwind-1.3.1/.configured_68b329da9893e34099c7d8ad5cb9c940] Error 1

Link: https://github.com/openwrt/packages/pull/12959#issuecomment-665021413
Reported-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>
Signed-off-by: Yousong Zhou <yszhou4tech@gmail.com>